ARM: mm: enforce pageblock alignment when freeing memmap entries at init
On some machines, the nodes do not always start on pageblock
boundaries. In these cases it is possible for free_unused_memmap
to free mappings for pages inside a pageblock with otherwise
valid pages. This presents problems for page migration since it
operates on whole pageblocks at a time.

Round down bank_start to pageblock boundary so that whole
pageblocks always have valid mappings.

Change-Id: Ie6588a274fc2fe9832cf47da31549ae3d004c118
Signed-off-by: Dima Zavin <dima@android.com>
Signed-off-by: KaSt <k@ka.st>