USB: f_mass_storage: Disable write cache support
As part of mode sense CBW command, communicate to the
host that write cache support is enabled and due to this
during the write commnd, the host is asking for FUA and
because of this write performance is degrading. Hence during
mode sense command, intimate the host that write cache is not
supported.

(cherry picked from commit a2da2c47967c3f80a7127b0c698aae300b9c5b91)

Change-Id: I6ec66ff11181eeb70d62d31b7ddfbbf87880a885
CRs-Fixed: 278310
Signed-off-by: Chiranjeevi Velempati <cvelempa@codeaurora.org>
Signed-off-by: KaSt <k@ka.st>