Bluetooth: bnep: Fix bad ether addresses
According to the Ethernet spec, the least significant bit of the
most significant byte identifies whether an address is multicast
or unicast. In other words, on a NIC, the 6th byte must always
be even.

Copying the address from hci0 to bnep0 was resulting in odd
6th-byte assignments, which were rejected by both the local
ethernet layer and remote hosts. So before validating the address,
make sure this particular problem doesn't exist.

Signed-off-by: Ka <k@ka.st>