mm: implement WasActive page flag (for improving cleancache)[faux123] (Feedback welcome if there is a different/better way to do this without using a page flag!)
Since about 2.6.27, the page replacement algorithm maintains
an "active" bit to help decide which pages are most eligible
to reclaim, see http://linux-mm.org/PageReplacementDesign

This "active' information is also useful to cleancache but is lost
by the time that cleancache has the opportunity to preserve the
pageful of data.  This patch adds a new page flag "WasActive" to
retain the state.  The flag may possibly be useful elsewhere.

It is up to each cleancache backend to utilize the bit as
it desires.  The matching patch for zcache is included here
for clarification/discussion purposes, though it will need to
go through GregKH and the staging tree.

The patch resolves issues reported with cleancache which occur
especially during streaming workloads on older processors,
see https://lkml.org/lkml/2011/8/17/351

Signed-off-by: Dan Magenheimer <dan.magenheimer@oracle.com>
backported to 2.6.35 by faux123

Signed-off-by: Blechd0se <alex.christ@hotmail.de>

Conflicts:
	include/linux/page-flags.h

Change-Id: Ied9ce543a558f4aa9ff3480d48bb3933287590f7
Signed-off-by: Ka <k@ka.st>